fix: handle no alias by adding display_prop conf
After my last refactor of alias handling, there was one case I didn't
consider:

What happens if no alias is defined, but you still want to
display a different window property than class as the title value?, and
match icons on that.

This fixes that by adding a new config value "display_property". It will
be checked when no alias is located, and return a title based of that.
Finally class is used, and if no properties are found it will return
with error, reporting in stdout.