fix: focus on old before renaming
Fixes issue with multi-monitor dragging would cause the workspace to
switch unexpectedly, possible fix for #34, but the reported issue is a
bit different than what I was able to reproduce.

Focus on the workspace about to be renamed before dispatching rename
commands ensures that the rename operation has a clear, unambiguous
target.

refs: #34 https://github.com/i3/i3/issues/5924